[noup] zephyr: fix 64-bit type mismatches in the Zephyr port
The Zephyr port has several places where a size_t value is passed through
an int-typed interface. These are invisible on 32-bit targets and only
surface when building for AArch64 or any other LP64 target.

In addition, hostapd_set_authorized() passes ~WPA_STA_AUTHORIZED to the
int-typed flags_and parameter of hostapd_sta_set_flags(). Because Zephyr
defines BIT() as (1UL << n), the expression is 64-bit on LP64 targets and
triggers a -Woverflow warning when converted to int. An explicit (int)
cast is added to make the intended truncation clear.

+++ wpa_supplicant/wpa_cli_zephyr.c
@@ -343,7 +343,7 @@ int zephyr_wpa_ctrl_zephyr_cmd(struct wpa_ctrl *ctrl, int argc, const char *argv
 int zephyr_wpa_cli_cmd_v(struct wpa_ctrl *ctrl, const char *fmt, ...)
 {
 	va_list cmd_args;
-	int argc;
+	size_t argc;
 	const char *argv[MAX_ARGS];
 	char cmd[MAX_CMD_SIZE];
 
@@ -353,9 +353,9 @@ int zephyr_wpa_cli_cmd_v(struct wpa_ctrl *ctrl, const char *fmt, ...)
 
 	(void)supp_make_argv(&argc, &argv[0], cmd, MAX_ARGS);
 
-	wpa_printf(MSG_DEBUG, "Calling wpa_cli: %s, argc: %d", cmd, argc);
-	for (int i = 0; i < argc; i++)
-		wpa_printf(MSG_DEBUG, "argv[%d]: %s", i, argv[i]);
+	wpa_printf(MSG_DEBUG, "Calling wpa_cli: %s, argc: %zu", cmd, argc);
+	for (size_t i = 0; i < argc; i++)
+		wpa_printf(MSG_DEBUG, "argv[%zu]: %s", i, argv[i]);
 
 	return zephyr_wpa_ctrl_zephyr_cmd(ctrl, argc, argv);
 }
@@ -429,7 +429,7 @@ int zephyr_wpa_global_ctrl_zephyr_cmd(int argc, const char *argv[])
 int zephyr_wpa_cli_global_cmd_v(const char *fmt, ...)
 {
 	va_list cmd_args;
-	int argc;
+	size_t argc;
 	const char *argv[MAX_ARGS] = {0};
 	char cmd[MAX_CMD_SIZE];
 
@@ -439,9 +439,9 @@ int zephyr_wpa_cli_global_cmd_v(const char *fmt, ...)
 
 	(void)supp_make_argv(&argc, &argv[0], cmd, MAX_ARGS);
 
-	wpa_printf(MSG_DEBUG, "Calling wpa_cli: %s, argc: %d", cmd, argc);
-	for (int i = 0; i < argc; i++)
-		wpa_printf(MSG_DEBUG, "argv[%d]: %s", i, argv[i]);
+	wpa_printf(MSG_DEBUG, "Calling wpa_cli: %s, argc: %zu", cmd, argc);
+	for (size_t i = 0; i < argc; i++)
+		wpa_printf(MSG_DEBUG, "argv[%zu]: %s", i, argv[i]);
 
 	return zephyr_wpa_global_ctrl_zephyr_cmd(argc, argv);
 }
